+++ Tests/ConfigTests/MergeTests.swift
@@ -8,49 +8,12 @@ var configDir: String {
     return path
 }
 
-class CLIConfigTests: XCTestCase {
-    func testCLI() throws {
-        let arguments = [
-            "--config:app.name=a",
-            "--config:app.friend.name=b",
-            "--config:app.friend.age=37",
-            "--config:name=world",
-            "--config:bools.yes"
-        ]
-        let cli = Node.makeCLIConfig(arguments: arguments)
-        let expectation: Node = [
-            "app": [
-                "name": "a",
-                "friend": [
-                    "name": "b",
-                    "age": "37"
-                ]
-            ],
-            "name": "world",
-            "bools": [
-                "yes": "true"
-            ]
-        ]
-        print(cli)
-        XCTAssertEqual(cli, expectation)
-    }
-
-    func testBools() {
-        let arguments = [
-            "--config:bools.yes"
-        ]
-        let cli = Node.makeCLIConfig(arguments: arguments)
-        let expectation: Node = [
-            "bools": [
-                "yes": "true"
-            ]
-        ]
-        print(cli)
-        XCTAssertEqual(cli, expectation)
-    }
-}
-
 class MergeTests: XCTestCase {
+    static let allTests = [
+        ("testFile", testFile),
+        ("testMerge", testMerge).
+    ]
+    
     func testFile() throws {
         let node = try Node.makeConfig(directory: configDir)
         let expectation: Node = [
